\name{isuR-package}
\alias{isuR-package}
\alias{isuR}
\docType{package}
\title{Introducing statistics using R}
\description{
  Mainly a collection of data sets for an introductionary lecture in statistics for natural scientists.
}
\details{
\tabular{ll}{
Package: \tab isuR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-29\cr
License: \tab What license is it under?\cr
}

}
\author{

Thomas Fabbro <thomas.fabbro@unibas.ch> 

}
\references{
no
}

\keyword{ package }
\examples{
library(help = "isuR")
}
