\name{stickleback}
\alias{stickleback}
\docType{data}
\title{
Body size of sticklebacks from the Lake Constance.
}
\description{
none
}
\usage{data(stickleback)}
\format{
  A data frame with 60 observations on the following 4 variables.
  \describe{
    \item{\code{individual}}{a factor with 60 levels indicating the
      different individuals measured}
    \item{\code{habitat}}{a factor with levels \code{inletStream} \code{lakeConstance}}
    \item{\code{sex}}{a factor with levels \code{f} \code{m}}
    \item{\code{bodySize}}{a numeric vector}
  }
}
\details{
none
}
\source{
Daniel Berner
}
\references{
none
}
\examples{
data(stickleback)
## maybe str(stickleback) ; plot(stickleback) ...
}
\keyword{datasets}
